worker/requirements.txt
pika
psycopg2-binary
redis
